backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
PyJWT==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6
pydantic[email]==2.6.0
python-dotenv==1.0.1